requirements.txt
libvirt-python==9.10.0
prometheus_client==0.20.0